Try layout for Sandbox, errors in WIndows VS17

+++ src/Combo/OpenGL/OpenGLBuffer.cpp
@@ -34,7 +34,7 @@ namespace Combo
     ----------------IndexBuffer-------------------
     --------------------------------------------*/
 
-    OpenGLIndexBuffer::OpenGLIndexBuffer(float* indices, uint32_t count)
+    OpenGLIndexBuffer::OpenGLIndexBuffer(uint32_t* indices, uint32_t count)
     : m_Count(count)
     {
         glCreateBuffers(1, &m_RendererID);
